+++ lib/pigeon_generator.dart
@@ -3,28 +3,23 @@ library;
 import 'package:build/build.dart';
 
 import 'src/pigeon_builder.dart';
-import 'src/pigeon_extensions.dart';
+import 'src/pigeon_config.dart';
+import 'src/pigeon_validator.dart';
 
+/// Creates a [PigeonBuilder] using the provided [BuilderOptions].
+///
+/// This function extracts the [PigeonConfig] from the [BuilderOptions],
+/// validates the configuration, and then returns a [PigeonBuilder] instance
+/// initialized with the validated configuration.
+///
+/// [options] - The [BuilderOptions] provided to the builder.
+///
+/// Returns a [PigeonBuilder] instance with the validated [PigeonConfig].
 Builder pigeonBuilder(BuilderOptions options) {
-  final config = options.config;
+  final pigeonConfig = PigeonConfig.fromBuilderOptions(options);
 
-  final platforms = (config['platforms'] as List? ?? []).toSet();
-
-  final platformOutputs = {
-    'cpp': platforms.contains('windows'),
-    'gobject': platforms.contains('linux'),
-    'kotlin': platforms.contains('android:kotlin'),
-    'java': platforms.contains('android:java'),
-    'swift': platforms.contains('ios'),
-    'objc': platforms.contains('macos'),
-  };
-
-  final pigeonConfig = PigeonExtension.loadConfig(
-    platformOutputs,
-    ast: config['ast'] ?? false,
-    dartTest: config['dart_test'] ?? false,
-    debugGenerators: config['debug_generators'] ?? false,
-  );
+  // Validate the pigeon config
+  PigeonValidator.validate(pigeonConfig);
 
   return PigeonBuilder(pigeonConfig);
 }