+++ lib/src/pigeon_builder.dart
@@ -8,15 +8,20 @@ import 'pigeon_config.dart';
 import 'pigeon_extensions.dart';
 import 'pigeon_scratch_space.dart';
 
+/// A [Builder] that uses Pigeon to generate code from Dart files.
 class PigeonBuilder extends Builder {
+  /// Creates a new [PigeonBuilder] with the given [PigeonConfig].
   PigeonBuilder(this.pigeonConfig);
 
+  /// The Pigeon configuration to use.
   final PigeonConfig pigeonConfig;
 
+  /// The path context to use for file operations.
   final path.Context _pathContext = path.Context(
     style: Platform.isWindows ? path.Style.posix : path.Style.platform,
   );
 
+  /// The resource for managing the Pigeon scratch space.
   final Resource<PigeonScratchSpace> _scratchSpaceResource = Resource(
     () => PigeonScratchSpace(),
     dispose: (scratchSpace) => scratchSpace.delete(),
@@ -80,36 +85,62 @@ class PigeonBuilder extends Builder {
     }
   }
 
+  /// Returns the [PigeonOptions] for the given input file.
   PigeonOptions _getPigeonOptions(String input) {
-    final fileName = path.basenameWithoutExtension(input);
+    String? getPath(
+      String? output,
+      String extension, {
+      String? append,
+      bool? capitalize,
+    }) {
+      if (output == null) return null;
 
-    String? getPath(String extension, String? output) {
-      return output != null ? '$output/$fileName$extension' : null;
+      final inputName = path.basenameWithoutExtension(input);
+
+      var outputName = inputName;
+
+      if (append != null) outputName += append;
+
+      if (capitalize == true) {
+        // If outputName is snake case, capitalize the first letter of each word
+        // and remove the underscores.
+        final regex = RegExp(r'(_[a-z])|(^[a-z])');
+        outputName = outputName.replaceAllMapped(regex, (Match match) {
+          return match[0]!.replaceAll('_', '').toUpperCase();
+        });
+      }
+
+      // Replace outTemplate name with outputName and extension with extension.
+      var outputFileName = pigeonConfig.outTemplate;
+      outputFileName = outputFileName.replaceAll('name', outputName);
+      outputFileName = outputFileName.replaceAll('extension', extension);
+
+      return path.join(output, outputFileName);
     }
 
     return PigeonOptions(
       input: input,
-      dartOut: getPath('.g.dart', pigeonConfig.dart?.out),
-      dartTestOut: getPath('_test.g.dart', pigeonConfig.dart?.testOut),
+      dartOut: getPath(pigeonConfig.dart?.out, 'dart'),
+      dartTestOut: getPath(pigeonConfig.dart?.testOut, 'dart', append: '_test'),
       dartPackageName: pigeonConfig.dart?.packageName,
-      cppHeaderOut: getPath('.g.h', pigeonConfig.cpp?.headerOut),
-      cppSourceOut: getPath('.g.cpp', pigeonConfig.cpp?.sourceOut),
+      cppHeaderOut: getPath(pigeonConfig.cpp?.headerOut, 'h'),
+      cppSourceOut: getPath(pigeonConfig.cpp?.sourceOut, 'cpp'),
       cppOptions: CppOptions(namespace: pigeonConfig.cpp?.namespace),
-      gobjectHeaderOut: getPath('.g.h', pigeonConfig.gobject?.headerOut),
-      gobjectSourceOut: getPath('.g.cc', pigeonConfig.gobject?.sourceOut),
+      gobjectHeaderOut: getPath(pigeonConfig.gobject?.headerOut, 'h'),
+      gobjectSourceOut: getPath(pigeonConfig.gobject?.sourceOut, 'cc'),
       gobjectOptions: GObjectOptions(module: pigeonConfig.gobject?.module),
-      kotlinOut: getPath('.g.kt', pigeonConfig.kotlin?.out),
+      kotlinOut: getPath(pigeonConfig.kotlin?.out, 'kt', capitalize: true),
       kotlinOptions: KotlinOptions(package: pigeonConfig.kotlin?.package),
-      javaOut: getPath('.g.java', pigeonConfig.java?.out),
+      javaOut: getPath(pigeonConfig.java?.out, 'java', capitalize: true),
       javaOptions: JavaOptions(
         package: pigeonConfig.java?.package,
         useGeneratedAnnotation: pigeonConfig.java?.useGeneratedAnnotation,
       ),
-      swiftOut: getPath('.g.swift', pigeonConfig.swift?.out),
-      objcHeaderOut: getPath('.g.h', pigeonConfig.objc?.headerOut),
-      objcSourceOut: getPath('.g.m', pigeonConfig.objc?.sourceOut),
+      swiftOut: getPath(pigeonConfig.swift?.out, 'swift', capitalize: true),
+      objcHeaderOut: getPath(pigeonConfig.objc?.headerOut, 'h'),
+      objcSourceOut: getPath(pigeonConfig.objc?.sourceOut, 'm'),
       objcOptions: ObjcOptions(prefix: pigeonConfig.objc?.prefix),
-      astOut: getPath('.g.ast', pigeonConfig.ast?.out),
+      astOut: getPath(pigeonConfig.ast?.out, 'ast'),
       debugGenerators: pigeonConfig.debugGenerators,
       copyrightHeader: pigeonConfig.copyrightHeader,
       oneLanguage: pigeonConfig.oneLanguage,