+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 2.0.0
+
+- Added validation checks for all outputs requiring them to be folders.
+- Added validation to check if copyright header file exists if provided in config.
+- Use out_template config to customize the generated file names.
+- Automatically generate the java/kotlin package folder structure based on applicationId in build.gradle.
+- All configurations are now done in build.yaml file instead of pigeon.yaml file.
+
 ## 1.0.0
 
 - Initial version.